% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyubishchev-package.R
\docType{package}
\name{lyubishchev-package}
\alias{lyubishchev}
\alias{lyubishchev-package}
\title{lyubishchev: Quantitative Taxonomy Methods of A.A. Lyubishchev (1943)}
\description{
Implements the multivariate classification methods of Alexander
Alexandrovich Lyubishchev (1890-1972), as described in his 1943 manuscript
\emph{Programma obshchey sistematiki} and published in Biometrics (1962).
}
\section{Main functions}{

\describe{
  \item{\code{\link{divergence_coefficient}}}{Standardised separation
    between two groups on continuous features.}
  \item{\code{\link{scatter_ellipse}}}{Fit covariance ellipses per class.}
  \item{\code{\link{transgression}}}{Detect overlap between two ellipses via
    Mahalanobis distance against a chi-squared threshold.}
  \item{\code{\link{classify}}}{Bayesian posterior classification of a new
    specimen.}
}

These methods predate and are more general than the binary-character
similarity coefficients of Sokal and Sneath (1963) that appear in other R
packages, operating directly on continuous Gaussian measurements.
}

\references{
Lyubishchev, A.A. (1943). Programma obshchey sistematiki
[Program of General Systematics]. Manuscript, 22 November 1943.
Digitized by ZIN RAS Coleoptera Laboratory.
\url{https://www.zin.ru/animalia/coleoptera/rus/lyubis05.htm}

Lubischew, A.A. (1962). On the use of discriminant functions in taxonomy.
Biometrics, 18(4), 455-477.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AkzhanBerdi/lyubishchev}
  \item Report bugs at \url{https://github.com/AkzhanBerdi/lyubishchev/issues}
}

}
\author{
\strong{Maintainer}: Akzhan Berdeyev \email{akzhan.berdeyev@gmail.com}

}
\keyword{internal}
